Move notification updates into a separate LifecycleObserver
